Move ReadSideTestDriver to tools project
