+++ tools/src/test/java/com/example/core/ReadSideTestDriver.java
@@ -1,4 +1,4 @@
-package com.example.auction.item.impl.testkit;
+package com.example.core;
 
 import akka.Done;
 import akka.japi.Pair;
@@ -6,7 +6,6 @@ import akka.stream.Materializer;
 import akka.stream.javadsl.Flow;
 import akka.stream.javadsl.Sink;
 import akka.stream.javadsl.Source;
-import com.example.core.CompletionStageUtils;
 import com.google.inject.Injector;
 import com.google.inject.Singleton;
 import com.lightbend.lagom.javadsl.persistence.*;
